requirements.txt
streamlit
google-generativeai
youtube-transcript-api
beautifulsoup4
requests